+++ demisto_sdk/commands/update_release_notes/update_rn.py
@@ -7,12 +7,16 @@ import os
 import re
 from distutils.version import LooseVersion
 from pathlib import Path
-from typing import Optional, Tuple, Union
+from typing import Any, Optional, Tuple, Union
 
 from demisto_sdk.commands.common.constants import (
     ALL_FILES_VALIDATION_IGNORE_WHITELIST, DEFAULT_ID_SET_PATH,
-    IGNORED_PACK_NAMES, RN_HEADER_BY_FILE_TYPE, FileType)
+    DEPRECATED_REGEXES, IGNORED_PACK_NAMES, RN_HEADER_BY_FILE_TYPE, FileType)
 from demisto_sdk.commands.common.content import Content
+from demisto_sdk.commands.common.content.objects.pack_objects import (
+    Integration, Playbook, Script)
+from demisto_sdk.commands.common.content.objects.pack_objects.abstract_pack_objects.yaml_content_object import \
+    YAMLContentObject
 from demisto_sdk.commands.common.git_util import GitUtil
 from demisto_sdk.commands.common.handlers import JSON_Handler
 from demisto_sdk.commands.common.tools import (LOG_COLORS, find_type,
@@ -28,6 +32,69 @@ from demisto_sdk.commands.common.tools import (LOG_COLORS, find_type,
                                                print_warning, run_command)
 
 json = JSON_Handler()
+
+CLASS_BY_FILE_TYPE = {FileType.INTEGRATION: Integration, FileType.SCRIPT: Script, FileType.PLAYBOOK: Playbook}
+
+
+def get_deprecated_comment_from_desc(description: str) -> str:
+    """
+    find deprecated comment from description
+    Args:
+        description: The yml description
+
+    Returns:
+        If a deprecated description is found return it for rn.
+    """
+    deprecate_line_with_replacement = re.findall(DEPRECATED_REGEXES[0], description)
+    deprecate_line_no_replacement = re.findall(DEPRECATED_REGEXES[1], description)
+
+    deprecate_line = deprecate_line_with_replacement + deprecate_line_no_replacement
+    return deprecate_line[0] if deprecate_line else ''
+
+
+def deprecated_commands(commands: list) -> set:
+    """return a set of the deprecated commands only"""
+    return {command.get('name') for command in commands if command.get('deprecated')}
+
+
+def get_yml_objects(path: str, file_type) -> Tuple[Any, YAMLContentObject]:
+    """
+    Generate yml files from master and from the current branch
+    Args:
+        path: The requested file path
+        file_type: the file type
+
+    Returns:
+        Two YML objects, the first of the yml at master (old yml) and the second from the current branch (new yml)
+    """
+    old_yml_obj = get_remote_file(path)
+    new_yml_obj = CLASS_BY_FILE_TYPE[file_type](path)
+
+    return old_yml_obj, new_yml_obj
+
+
+def get_deprecated_rn(path: str, file_type):
+    """Generate rn for deprecated items"""
+    old_yml, new_yml = get_yml_objects(path, file_type)
+
+    if not old_yml.get('deprecated') and new_yml.is_deprecated:
+        description = new_yml.get('comment') if file_type == file_type.SCRIPT else new_yml.get('description')
+        rn_from_description = get_deprecated_comment_from_desc(description)
+        return f'- Deprecated. {rn_from_description or "Use %%% instead"}.\n'
+
+    if file_type != FileType.INTEGRATION:
+        return ''
+
+    # look for deprecated commands
+    rn = ''
+    old_commands = deprecated_commands(old_yml.get('script', {}).get('commands'))
+    new_commands = deprecated_commands(new_yml.script.get('commands'))
+
+    for command_name in new_commands:
+        # if command is deprecated in new yml, and not in old yml
+        if command_name not in old_commands:
+            rn += f'- Command ***{command_name}*** is deprecated. Use %%% instead.\n'
+    return rn
 
 
 class UpdateRN:
@@ -548,7 +615,15 @@ class UpdateRN:
                 if self.update_type == 'documentation':
                     rn_desc += '- Documentation and metadata improvements.\n'
                 else:
-                    rn_desc += f'- {text or "%%UPDATE_RN%%"}\n'
+                    deprecate_rn = ''
+                    if _type in (FileType.INTEGRATION, FileType.SCRIPT, FileType.PLAYBOOK):
+                        deprecate_rn = get_deprecated_rn(path, _type)
+                    if deprecate_rn:
+                        if text:
+                            rn_desc += f'- {text}\n'
+                        rn_desc += deprecate_rn
+                    else:
+                        rn_desc += f'- {text or "%%UPDATE_RN%%"}\n'
 
         if _type == FileType.TRIGGER:
             rn_desc = f'- {desc}'  # Issue - https://github.com/demisto/etc/issues/48153#issuecomment-1111988526